Fixed jtw filter
