--- src/main/java/com/example/gateway_api/filters/JwtAuthenticationFilter.java
@@ -1,76 +1,0 @@
-package com.example.gateway_api.filters;
-
-import org.springframework.cloud.gateway.filter.GatewayFilter;
-import org.springframework.cloud.gateway.filter.GatewayFilterChain;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.server.reactive.ServerHttpRequest;
-import org.springframework.stereotype.Component;
-import org.springframework.web.reactive.function.client.WebClient;
-import org.springframework.web.server.ServerWebExchange;
-import reactor.core.publisher.Mono;
-
-@Component
-public class JwtAuthenticationFilter implements GatewayFilter {
-
-    private final WebClient webClient;
-
-    public JwtAuthenticationFilter(WebClient.Builder webClientBuilder) {
-        this.webClient = webClientBuilder.baseUrl("http://auth-api:8005/auth").build();
-    }
-
-    @Override
-    public Mono<Void> filter(ServerWebExchange exchange, GatewayFilterChain chain) {
-        String path = exchange.getRequest().getURI().getPath();
-
-        // Bypass authentication for public endpoints
-        if (path.equals("/auth/signup") || path.equals("/auth/login")) {
-            return chain.filter(exchange);
-        }
-
-        String authHeader = exchange.getRequest().getHeaders().getFirst(HttpHeaders.AUTHORIZATION);
-
-        if (authHeader == null || !authHeader.startsWith("Bearer ")) {
-            exchange.getResponse().setStatusCode(HttpStatus.UNAUTHORIZED);
-            return exchange.getResponse().setComplete();
-        }
-
-        String token = authHeader.substring(7);
-
-        return webClient.post()
-                .uri("/verify") // Endpoint on auth service
-                .header(HttpHeaders.AUTHORIZATION, authHeader)
-                .retrieve()
-                .bodyToMono(VerifiedTokenResponse.class)
-                .flatMap(response -> {
-                    // Add user information as headers for downstream services
-                    ServerHttpRequest mutatedRequest = exchange.getRequest().mutate()
-                            .header("X-User-Id", response.getUserId())
-                            .header("X-User-Role", response.getUserRole())
-                            .build();
-                    return chain.filter(exchange.mutate().request(mutatedRequest).build());
-                });
-    }
-
-    public static class VerifiedTokenResponse {
-        private String userId;
-        private String userRole;
-
-        public String getUserId() {
-            return userId;
-        }
-
-        public void setUserId(String userId) {
-            this.userId = userId;
-        }
-
-        public String getUserRole() {
-            return userRole;
-        }
-
-        public void setUserRole(String userRole) {
-            this.userRole = userRole;
-        }
-    }
-}
-
