+++ src/main/resources/application.properties
@@ -1,5 +1,4 @@
 spring.application.name=gateway-api
-
 #logging.level.org.springframework.web.cors.reactive=DEBUG:
 #logging.level.org.springframework.web: DEBUG
 #logging.level.org.springframework.security: DEBUG
